# api/openapi.yaml
openapi: 3.0.3
info:
  title: CEX Router API
  version: 0.1.0
  description: Public read API for centralized-exchange deposit and withdrawal rail status.
  license:
    name: Not specified
    url: https://choosealicense.com/no-permission/
servers:
  - url: /
security: []
...
paths:
...
  /v1/adapters/freshness:
    get:
      tags:
        - System
      summary: List adapter freshness for each exchange
      operationId: listAdapterFreshness
      responses:
        "200":
          description: Adapter polling freshness rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterFreshnessListResponse"
        "400":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
components:
...
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
    Exchange:
      type: object
      required:
        - id
        - slug
        - name
        - region
      properties:
        id:
          type: integer
          format: int32
        slug:
          type: string
        name:
          type: string
        region:
          type: string
...
    AdapterFreshness:
      type: object
      required:
        - exchange
        - last_successful_poll
        - last_attempt
        - last_error
        - consecutive_failures
      properties:
        exchange:
          $ref: "#/components/schemas/Exchange"
        last_successful_poll:
          type: string
          format: date-time
          nullable: true
        last_attempt:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
        consecutive_failures:
          type: integer
          format: int32
...
    AdapterFreshnessListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdapterFreshness"